pyproject.toml
[project]
name = "fl_chan"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "NingmengLemon", email = "60469383+NingmengLemon@users.noreply.github.com"},
]
dependencies = ["pillow>=11.1.0"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
